Fixed the `insertHead` & `insertAt` functions
Previously function `insertHead` did not pass the correct `next` pointer
to the new head resulting in ommiting the former head after insertion.
In the result size was incremented by one, but the actual size of
the linked list did not change.

On the other hand `insertAt` shall not be called when pos is the same
as the current size of the list. The insertion procedure shall begin
at the `pos - 1` index.

Also fixed some traces to be more readable.