Minor cleanup of getBucketEntry.
Method getBuceketEntry caused an undefined behavior once used. Starting from
invalid alloceted block read-size access, to literally throwing 2 errors
depending on the test seed run:
1. std::bad_alloc
2. basic_string::_M_create - std::lenght_error
I am not sure why this error occured on the first place but it's fixed now.
It's safe to use this function as it will now directly operate on a table_
instead of using a copy.